requirements.txt
duckdb>=1.1.3
requests>=2.32.3
datafusion>=43.1.0
python-dotenv>=1.1.0